+++ twitter4j-core/src/main/java/twitter4j/internal/json/StatusJSONImpl.java
@@ -47,7 +47,7 @@ import static twitter4j.internal.json.z_T4JInternalParseUtil.*;
     private long inReplyToUserId;
     private boolean isFavorited;
     private boolean isRetweeted;
-    private long favoriteCount;
+    private int favoriteCount;
     private String inReplyToScreenName;
     private GeoLocation geoLocation = null;
     private Place place = null;
@@ -104,7 +104,7 @@ import static twitter4j.internal.json.z_T4JInternalParseUtil.*;
         isRetweeted = getBoolean("retweeted", json);
         inReplyToScreenName = getUnescapedString("in_reply_to_screen_name", json);
         retweetCount = getLong("retweet_count", json);
-        favoriteCount = getLong("favorite_count", json);
+        favoriteCount = getInt("favorite_count", json);
         isPossiblySensitive = getBoolean("possibly_sensitive", json);
         try {
             if (!json.isNull("user")) {
@@ -312,12 +312,12 @@ import static twitter4j.internal.json.z_T4JInternalParseUtil.*;
     public boolean isRetweeted() {
         return isRetweeted;
     }
-    
+
     /**
      * {@inheritDoc}
      */
     @Override
-    public long getFavoriteCount() {
+    public int getFavoriteCount() {
         return favoriteCount;
     }
